Multi-chain hosting for ao-recorder
Prompt: "continue with most logical next steps" — implement multi-chain hosting,
the last remaining Phase 2 code feature.

- Refactor AppState into per-chain ChainState (store, blockmaker_key, broadcast)
  plus shared AppState holding RwLock<HashMap<chain_id, Arc<ChainState>>>
- Add GET /chains (list all hosted chains) and POST /chains (create chain from
  genesis JSON at runtime) endpoints
- All existing /chain/{id}/... handlers now look up chain by ID from the map,
  returning 404 for unknown chains
- Config supports [[chains]] array for loading multiple chains at startup, plus
  optional data_dir for file-backed dynamic chain creation
- Backward-compatible: AppState::new(store, key) still works for single-chain
- 4 new integration tests: list_chains, create_chain_via_api,
  create_duplicate_chain_rejected, multi_chain_independent_operations
- Fix new clippy lints from toolchain update: is_multiple_of, collapsible_if,
  print_literal, needless_lifetimes, unnecessary_cast
- 86 tests pass, zero warnings

+++ 2026/src/ao-chain/src/store.rs
@@ -395,7 +395,7 @@ impl ChainStore {
     pub fn store_block(&self, height: u64, timestamp: i64, hash: &[u8; 32], data: &[u8]) -> Result<()> {
         self.conn.execute(
             "INSERT INTO blocks (height, timestamp, hash, data) VALUES (?1, ?2, ?3, ?4)",
-            params![height as i64, timestamp as i64, &hash[..], data],
+            params![height as i64, timestamp, &hash[..], data],
         )?;
         Ok(())
     }
